pyproject.toml
[project]
name = "latentscience"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.11",
    "aiosignal>=1.3.2",
    "alembic>=1.16.1",
    "annotated-types>=0.7.0",
    "anthropic>=0.52.2",
    "anyio>=4.9.0",
    "attrs>=25.3.0",
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "blis>=1.3.0",
    "catalogue>=2.0.10",
    "certifi>=2025.4.26",
    "charset-normalizer>=3.4.2",
    "click>=8.1.8",
    "cloudpathlib>=0.21.1",
    "confection>=0.1.5",
    "cymem>=2.0.11",
    "dishka>=1.6.0",
    "distro>=1.9.0",
    "fastapi>=0.115.12",
    "filelock>=3.18.0",
    "flake8>=7.2.0",
    "frozenlist>=1.6.2",
    "fsspec>=2025.5.1",
    "grpclib>=0.4.7",
    "h11>=0.16.0",
    "h2>=4.2.0",
    "hf-xet>=1.1.3",
    "hpack>=4.1.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "huggingface-hub>=0.32.4",
    "hyperframe>=6.1.0",
    "idna>=3.10",
    "iniconfig>=2.1.0",
    "jinja2>=3.1.6",
    "jiter>=0.10.0",
    "joblib>=1.5.1",
    "langcodes>=3.5.0",
    "language-data>=1.3.0",
    "mako>=1.3.10",
    "marisa-trie>=1.2.1",
    "markdown-it-py>=3.0.0",
    "markupsafe>=3.0.2",
    "mccabe>=0.7.0",
    "mdurl>=0.1.2",
    "modal>=1.0.3",
    "mpmath>=1.3.0",
    "multidict>=6.4.4",
    "murmurhash>=1.0.13",
    "mypy>=1.16.0",
    "mypy-extensions>=1.1.0",
    "networkx>=3.5",
    "nltk>=3.9.1",
    "numpy>=2.3.0",
    "openai>=1.84.0",
    "packaging>=25.0",
    "pathspec>=0.12.1",
    "pillow>=11.2.1",
    "platformdirs>=4.3.8",
    "pluggy>=1.6.0",
    "preshed>=3.0.10",
    "propcache>=0.3.1",
    "protobuf>=6.31.1",
    "psycopg2-binary>=2.9.10",
    "pycodestyle>=2.13.0",
    "pydantic>=2.11.5",
    "pydantic-core>=2.33.2",
    "pydantic-settings>=2.9.1",
    "pyflakes>=3.3.2",
    "pygments>=2.19.1",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "regex>=2024.11.6",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "safetensors>=0.5.3",
    "scikit-learn>=1.7.0",
    "scipy>=1.15.3",
    "sentence-transformers>=4.1.0",
    "setuptools>=80.9.0",
    "shellingham>=1.5.4",
    "sigtools>=4.0.1",
    "smart-open>=7.1.0",
    "sniffio>=1.3.1",
    "soupsieve>=2.7",
    "spacy>=3.8.7",
    "spacy-legacy>=3.0.12",
    "spacy-loggers>=1.0.5",
    "srsly>=2.5.1",
    "starlette>=0.46.2",
    "sympy>=1.14.0",
    "synchronicity>=0.9.13",
    "thinc>=8.3.6",
    "threadpoolctl>=3.6.0",
    "tokenizers>=0.21.1",
    "toml>=0.10.2",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
    "typer>=0.16.0",
    "types-certifi>=2021.10.8.3",
    "types-toml>=0.10.8.20240310",
    "typing-extensions>=4.14.0",
    "typing-inspection>=0.4.1",
    "urllib3>=2.4.0",
    "uvicorn>=0.34.3",
    "wasabi>=1.1.3",
    "watchfiles>=1.0.5",
    "weasel>=0.4.1",
    "wrapt>=1.17.2",
    "yarl>=1.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["latentscience"]
